--- react/react-app/src/components/cart/breadcrumb/breadcrumbPart.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const BreadcrumbPart = (props) =>{
-    return(
-        <a href="/" className = "breadcrumb">{props.breadcrumbNav}</a>
-    )
-}
-
-export  default BreadcrumbPart;
