--- react/react-app/src/components/cart/cartTitle/carttitle.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import Title from './title';
-
-const CartTitle = (props) =>{
-    return(
-        <Title cartTitle ="Cart" />
-    )
-};
-
-export default CartTitle;
