--- react/react-app/src/components/cart/cartTitle/title.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Title = (props) =>{
-    return(
-        <h3 className ='title'>{props.cartTitle}</h3>
-    )
-};
-
-export default Title;
