--- react/react-app/src/components/verificationpage/verfiBreadcrumb.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import BreadcrumbPart from '../cart/breadcrumb/breadcrumbPart';
-
-const VerifBreadcrumb =(props)=>{
-    return(
-        <div>
-            <BreadcrumbPart breadcrumbNav ="Home >"/>
-            <BreadcrumbPart breadcrumbNav ="Cart >"/>
-            <BreadcrumbPart breadcrumbNav ="Verification"/>
-        </div>
-    )
-};
-export default VerifBreadcrumb;
