+++ submissions/mariamPopoola/react/react/react-app/src/App.js
@@ -1,0 +1,15 @@
+import './App.css';
+import Verification from '../src/components/verificationpage/verification';
+
+function App() {
+  return (
+    <div>
+       <Verification/>
+    </div>
+  );
+}
+
+export default App;
+
+  
+
