+++ submissions/mariamPopoola/react/react/react-app/src/components/cart/breadcrumb/breadcrumb.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import BreadcrumbPart from './breadcrumbPart';
+
+const Breadcrumb = () =>{
+    return(
+        <div>
+            <BreadcrumbPart breadcrumbNav = "Home > "/>
+            <BreadcrumbPart breadcrumbNav = "cart"/>
+        </div>
+    )
+};
+export default Breadcrumb;
