+++ submissions/mariamPopoola/react/react/react-app/src/components/sucesspage/success.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import Header  from '../header'
+import SuccessPart from './sucesspart'
+import success from '../../images/success.svg'
+
+const Success =(props)=>{
+    return(
+        <div>
+            <Header/>
+            <SuccessPart 
+            image = {success}
+            imageName = "sucess"
+            order ="Order successfull"
+            orderMessage="We have started processing your order"
+            />
+        </div>
+      
+    )
+};
+export default Success;
+
