+++ submissions/mariamPopoola/react/react/react-app/src/components/verificationpage/form/button.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+function Button(props){
+    return(
+            <button className ="" type={props.inputType}>
+                {props.title}
+            </button>
+    )
+}
+export default Button;
