+++ submissions/mariamPopoola/react/react/react-app/src/components/verificationpage/verification.js
@@ -1,0 +1,13 @@
+import React  from "react";
+import Header from '../header'
+import VerificationPart from "./verificationpart";
+
+const Verification =(props)=> {
+    return(
+        <div>
+            <Header/>
+            <VerificationPart/>
+        </div>
+    )
+};
+export default Verification;
