+++ submissions/mariamPopoola/react/react/react-app/src/index.css
@@ -1,0 +1,11 @@
+*{
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Poppins', sans-serif;
+}
+
+body{
+  padding-right: 30px;
+  padding-left: 30px;
+}
